Splitted the AutomationElement in separate files for AsMethods and Find
