+++ src/FlaUI.Core/AutomationElements/Infrastructure/AutomationElement.cs
@@ -1,17 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
-using FlaUI.Core.AutomationElements.Scrolling;
 using FlaUI.Core.Conditions;
 using FlaUI.Core.Definitions;
 using FlaUI.Core.EventHandlers;
 using FlaUI.Core.Exceptions;
 using FlaUI.Core.Identifiers;
 using FlaUI.Core.Input;
-using FlaUI.Core.Tools;
 using FlaUI.Core.WindowsAPI;
 using GdiColor = System.Drawing.Color;
 using Rectangle = FlaUI.Core.Shapes.Rectangle;
@@ -22,7 +19,7 @@ namespace FlaUI.Core.AutomationElements.Infrastructure
     /// <summary>
     /// Wrapper object for each ui element which is automatable.
     /// </summary>
-    public class AutomationElement : IEquatable<AutomationElement>
+    public partial class AutomationElement : IEquatable<AutomationElement>
     {
         /// <summary>
         /// Creates a new instance which wraps around the given <see cref="BasicAutomationElement"/>.
@@ -347,123 +344,6 @@ namespace FlaUI.Core.AutomationElements.Infrastructure
         }
 
         /// <summary>
-        /// Finds all elements in the given treescope and with the given condition.
-        /// </summary>
-        public AutomationElement[] FindAll(TreeScope treeScope, ConditionBase condition)
-        {
-            return FindAll(treeScope, condition, Retry.DefaultRetryFor);
-        }
-
-        /// <summary>
-        /// Finds all elements in the given treescope and with the given condition within the given timeout.
-        /// </summary>
-        public AutomationElement[] FindAll(TreeScope treeScope, ConditionBase condition, TimeSpan timeOut)
-        {
-            Predicate<AutomationElement[]> whilePredicate = elements => elements.Length == 0;
-            Func<AutomationElement[]> retryMethod = () => BasicAutomationElement.FindAll(treeScope, condition);
-            return Retry.While(retryMethod, whilePredicate, timeOut);
-        }
-
-        /// <summary>
-        /// Finds the first element which is in the given treescope with the given condition.
-        /// </summary>
-        public AutomationElement FindFirst(TreeScope treeScope, ConditionBase condition)
-        {
-            return FindFirst(treeScope, condition, Retry.DefaultRetryFor);
-        }
-
-        /// <summary>
-        /// Finds the first element which is in the given treescope with the given condition within the given timeout period.
-        /// </summary>
-        public AutomationElement FindFirst(TreeScope treeScope, ConditionBase condition, TimeSpan timeOut)
-        {
-            Predicate<AutomationElement> whilePredicate = element => element == null;
-            Func<AutomationElement> retryMethod = () => BasicAutomationElement.FindFirst(treeScope, condition);
-            return Retry.While(retryMethod, whilePredicate, timeOut);
-        }
-
-        /// <summary>
-        /// Finds the first element by iterating thru all conditions.
-        /// </summary>
-        public AutomationElement FindFirstNested(params ConditionBase[] nestedConditions)
-        {
-            var currentElement = this;
-            foreach (var condition in nestedConditions)
-            {
-                currentElement = currentElement.FindFirstChild(condition);
-                if (currentElement == null)
-                {
-                    return null;
-                }
-            }
-            return currentElement;
-        }
-
-        /// <summary>
-        /// Finds all elements by iterating thru all conditions.
-        /// </summary>
-        public AutomationElement[] FindAllNested(params ConditionBase[] nestedConditions)
-        {
-            var currentElement = this;
-            for (var i = 0; i < nestedConditions.Length - 1; i++)
-            {
-                var condition = nestedConditions[i];
-                currentElement = currentElement.FindFirstChild(condition);
-                if (currentElement == null)
-                {
-                    return null;
-                }
-            }
-            return currentElement.FindAllChildren(nestedConditions.Last());
-        }
-
-        /// <summary>
-        /// Finds for the first item which matches the given xpath.
-        /// </summary>
-        public AutomationElement FindFirstByXPath(string xPath)
-        {
-            var xPathNavigator = new AutomationElementXPathNavigator(this);
-            var nodeItem = xPathNavigator.SelectSingleNode(xPath);
-            return (AutomationElement)nodeItem?.UnderlyingObject;
-        }
-
-        /// <summary>
-        /// Finds all items which match the given xpath.
-        /// </summary>
-        public AutomationElement[] FindAllByXPath(string xPath)
-        {
-            var xPathNavigator = new AutomationElementXPathNavigator(this);
-            var itemNodeIterator = xPathNavigator.Select(xPath);
-            var itemList = new List<AutomationElement>();
-            while (itemNodeIterator.MoveNext())
-            {
-                var automationItem = (AutomationElement)itemNodeIterator.Current.UnderlyingObject;
-                itemList.Add(automationItem);
-            }
-            return itemList.ToArray();
-        }
-
-        /// <summary>
-        /// Finds the element which is in the given treescope with the given condition and the given index.
-        /// </summary>
-        public AutomationElement FindAt(TreeScope treeScope, int index, ConditionBase condition)
-        {
-            Predicate<AutomationElement> whilePredicate = element => element == null;
-            Func<AutomationElement> retryMethod = () => BasicAutomationElement.FindIndexed(treeScope, index, condition);
-            return Retry.While(retryMethod, whilePredicate, Retry.DefaultRetryFor);
-        }
-
-        /// <summary>
-        /// Finds the element which is in the given treescope with the given condition and the given index within the given timeout period.
-        /// </summary>
-        public AutomationElement FindAt(TreeScope treeScope, ConditionBase condition, int index, TimeSpan timeOut)
-        {
-            Predicate<AutomationElement> whilePredicate = element => element == null;
-            Func<AutomationElement> retryMethod = () => BasicAutomationElement.FindIndexed(treeScope, index, condition);
-            return Retry.While(retryMethod, whilePredicate, timeOut);
-        }
-
-        /// <summary>
         /// Gets a clickable point of the element.
         /// </summary>
         /// <exception cref="Exceptions.NoClickablePointException">Thrown when no clickable point was found</exception>
@@ -671,429 +551,6 @@ namespace FlaUI.Core.AutomationElements.Infrastructure
         protected virtual void SetFocus()
         {
             BasicAutomationElement.SetFocus();
-        } 
-
-        #region Conversion Methods
-        /// <summary>
-        /// Converts the element to a <see cref="Button"/>.
-        /// </summary>
-        public Button AsButton()
-        {
-            return new Button(BasicAutomationElement);
         }
-
-        /// <summary>
-        /// Converts the element to a <see cref="CheckBox"/>.
-        /// </summary>
-        public CheckBox AsCheckBox()
-        {
-            return new CheckBox(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="ComboBox"/>.
-        /// </summary>
-        public ComboBox AsComboBox()
-        {
-            return new ComboBox(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="DataGridView"/>.
-        /// </summary>
-        public DataGridView AsDataGridView()
-        {
-            return new DataGridView(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Label"/>.
-        /// </summary>
-        public Label AsLabel()
-        {
-            return new Label(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Grid"/>.
-        /// </summary>
-        public Grid AsGrid()
-        {
-            return new Grid(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="GridRow"/>.
-        /// </summary>
-        public GridRow AsGridRow()
-        {
-            return new GridRow(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="GridCell"/>.
-        /// </summary>
-        public GridCell AsGridCell()
-        {
-            return new GridCell(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="GridHeader"/>.
-        /// </summary>
-        public GridHeader AsGridHeader()
-        {
-            return new GridHeader(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="GridHeaderItem"/>.
-        /// </summary>
-        public GridHeaderItem AsGridHeaderItem()
-        {
-            return new GridHeaderItem(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="HorizontalScrollBar"/>.
-        /// </summary>
-        public HorizontalScrollBar AsHorizontalScrollBar()
-        {
-            return new HorizontalScrollBar(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="ListBox"/>.
-        /// </summary>
-        public ListBox AsListBox()
-        {
-            return new ListBox(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="ListBoxItem"/>.
-        /// </summary>
-        public ListBoxItem AsListBoxItem()
-        {
-            return new ListBoxItem(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Menu"/>.
-        /// </summary>
-        public Menu AsMenu()
-        {
-            return new Menu(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="MenuItem"/>.
-        /// </summary>
-        public MenuItem AsMenuItem()
-        {
-            return new MenuItem(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="ProgressBar"/>.
-        /// </summary>
-        public ProgressBar AsProgressBar()
-        {
-            return new ProgressBar(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="RadioButton"/>.
-        /// </summary>
-        public RadioButton AsRadioButton()
-        {
-            return new RadioButton(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Slider"/>.
-        /// </summary>
-        public Slider AsSlider()
-        {
-            return new Slider(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Tab"/>.
-        /// </summary>
-        public Tab AsTab()
-        {
-            return new Tab(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="TabItem"/>.
-        /// </summary>
-        public TabItem AsTabItem()
-        {
-            return new TabItem(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="TextBox"/>.
-        /// </summary>
-        public TextBox AsTextBox()
-        {
-            return new TextBox(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Thumb"/>.
-        /// </summary>
-        public Thumb AsThumb()
-        {
-            return new Thumb(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="TitleBar"/>.
-        /// </summary>
-        public TitleBar AsTitleBar()
-        {
-            return new TitleBar(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="ToggleButton"/>.
-        /// </summary>
-        public ToggleButton AsToggleButton()
-        {
-            return new ToggleButton(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Tree"/>.
-        /// </summary>
-        public Tree AsTree()
-        {
-            return new Tree(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="TreeItem"/>.
-        /// </summary>
-        public TreeItem AsTreeItem()
-        {
-            return new TreeItem(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="VerticalScrollBar"/>.
-        /// </summary>
-        public VerticalScrollBar AsVerticalScrollBar()
-        {
-            return new VerticalScrollBar(BasicAutomationElement);
-        }
-
-        /// <summary>
-        /// Converts the element to a <see cref="Window"/>.
-        /// </summary>
-        public Window AsWindow()
-        {
-            return new Window(BasicAutomationElement);
-        }
-        #endregion Conversion Methods
-
-        #region Convenience methods
-        /// <summary>
-        /// Finds the first child.
-        /// </summary>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstChild()
-        {
-            return FindFirst(TreeScope.Children, TrueCondition.Default);
-        }
-
-        /// <summary>
-        /// Finds the first child with the given automation id.
-        /// </summary>
-        /// <param name="automationId">The automation id.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstChild(string automationId)
-        {
-            return FindFirst(TreeScope.Children, ConditionFactory.ByAutomationId(automationId));
-        }
-
-        /// <summary>
-        /// Finds the first child with the condition.
-        /// </summary>
-        /// <param name="condition">The condition.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstChild(ConditionBase condition)
-        {
-            return FindFirst(TreeScope.Children, condition);
-        }
-
-        /// <summary>
-        /// Finds the first child with the condition.
-        /// </summary>
-        /// <param name="conditionFunc">The condition method.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstChild(Func<ConditionFactory, ConditionBase> conditionFunc)
-        {
-            var condition = conditionFunc(ConditionFactory);
-            return FindFirstChild(condition);
-        }
-
-        /// <summary>
-        /// Finds all children.
-        /// </summary>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllChildren()
-        {
-            return FindAll(TreeScope.Children, TrueCondition.Default);
-        }
-
-        /// <summary>
-        /// Finds all children with the condition.
-        /// </summary>
-        /// <param name="condition">The condition.</param>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllChildren(ConditionBase condition)
-        {
-            return FindAll(TreeScope.Children, condition);
-        }
-
-        /// <summary>
-        /// Finds all children with the condition.
-        /// </summary>
-        /// <param name="conditionFunc">The condition mehtod.</param>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllChildren(Func<ConditionFactory, ConditionBase> conditionFunc)
-        {
-            var condition = conditionFunc(ConditionFactory);
-            return FindAllChildren(condition);
-        }
-
-        /// <summary>
-        /// Finds the first descendant.
-        /// </summary>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstDescendant()
-        {
-            return FindFirst(TreeScope.Descendants, TrueCondition.Default);
-        }
-
-        /// <summary>
-        /// Finds the first descendant with the given automation id.
-        /// </summary>
-        /// <param name="automationId">The automation id.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstDescendant(string automationId)
-        {
-            return FindFirst(TreeScope.Descendants, ConditionFactory.ByAutomationId(automationId));
-        }
-
-        /// <summary>
-        /// Finds the first descendant with the condition.
-        /// </summary>
-        /// <param name="condition">The condition.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstDescendant(ConditionBase condition)
-        {
-            return FindFirst(TreeScope.Descendants, condition);
-        }
-
-        /// <summary>
-        /// Finds the first descendant with the condition.
-        /// </summary>
-        /// <param name="conditionFunc">The condition method.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstDescendant(Func<ConditionFactory, ConditionBase> conditionFunc)
-        {
-            var condition = conditionFunc(ConditionFactory);
-            return FindFirstDescendant(condition);
-        }
-
-        /// <summary>
-        /// Finds all descendants.
-        /// </summary>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllDescendants()
-        {
-            return FindAll(TreeScope.Descendants, TrueCondition.Default);
-        }
-
-        /// <summary>
-        /// Finds all descendants with the condition.
-        /// </summary>
-        /// <param name="condition">The condition.</param>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllDescendants(ConditionBase condition)
-        {
-            return FindAll(TreeScope.Descendants, condition);
-        }
-
-        /// <summary>
-        /// Finds all descendants with the condition.
-        /// </summary>
-        /// <param name="conditionFunc">The condition mehtod.</param>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllDescendants(Func<ConditionFactory, ConditionBase> conditionFunc)
-        {
-            var condition = conditionFunc(ConditionFactory);
-            return FindAllDescendants(condition);
-        }
-
-        /// <summary>
-        /// Finds the first element by iterating thru all conditions.
-        /// </summary>
-        /// <param name="conditionFunc">The condition method.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindFirstNested(Func<ConditionFactory, IList<ConditionBase>> conditionFunc)
-        {
-            var conditions = conditionFunc(ConditionFactory);
-            return FindFirstNested(conditions.ToArray());
-        }
-
-        /// <summary>
-        /// Finds all elements by iterating thru all conditions.
-        /// </summary>
-        /// <param name="conditionFunc">The condition method.</param>
-        /// <returns>The found elements or an empty list if no elements were found.</returns>
-        public AutomationElement[] FindAllNested(Func<ConditionFactory, IList<ConditionBase>> conditionFunc)
-        {
-            var conditions = conditionFunc(ConditionFactory);
-            return FindAllNested(conditions.ToArray());
-        }
-
-        /// <summary>
-        /// Finds the child at the given position.
-        /// </summary>
-        /// <param name="index">The index of the child to find.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindChildAt(int index)
-        {
-            return FindChildAt(index, TrueCondition.Default);
-        }
-
-        /// <summary>
-        /// Finds the child at the given position with the condition.
-        /// </summary>
-        /// <param name="index">The index of the child to find.</param>
-        /// <param name="condition">The condition.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindChildAt(int index, ConditionBase condition)
-        {
-            return FindAt(TreeScope.Children, index, condition);
-        }
-
-        /// <summary>
-        /// Finds the child at the given position with the condition.
-        /// </summary>
-        /// <param name="index">The index of the child to find.</param>
-        /// <param name="conditionFunc">The condition mehtod.</param>
-        /// <returns>The found element or null if no element was found.</returns>
-        public AutomationElement FindChildAt(int index, Func<ConditionFactory, ConditionBase> conditionFunc)
-        {
-            var condition = conditionFunc(ConditionFactory);
-            return FindChildAt(index, condition);
-        }
-        #endregion Convenience methods
     }
 }
